Fix issue with truncate table and switching memory contexts - Moved ExecuteWithMpMemory and related classes to separate file - Added new class ExecuteWithAllSitesMemory - Made static variable s_enginesByPartitionId private - Removed references to DEBUG, as NDEBUG is standard
