+++ src/ee/common/ExecuteWithMpMemory.cpp
@@ -1,0 +1,111 @@
+/* This file is part of VoltDB.
+ * Copyright (C) 2008-2018 VoltDB Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with VoltDB.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "ExecuteWithMpMemory.h"
+
+#include "common/debuglog.h"
+#include "common/SynchronizedThreadLock.h"
+
+namespace voltdb {
+
+ExecuteWithMpMemory::ExecuteWithMpMemory() {
+    VOLT_DEBUG("Entering UseMPmemory");
+    SynchronizedThreadLock::assumeMpMemoryContext();
+}
+ExecuteWithMpMemory::~ExecuteWithMpMemory() {
+    VOLT_DEBUG("Exiting UseMPmemory");
+    SynchronizedThreadLock::assumeLocalSiteContext();
+}
+
+ConditionalExecuteWithMpMemory::ConditionalExecuteWithMpMemory(bool needMpMemory) : m_usingMpMemory(needMpMemory) {
+    if (m_usingMpMemory) {
+        VOLT_DEBUG("Entering UseMPmemory");
+        SynchronizedThreadLock::assumeMpMemoryContext();
+    }
+}
+
+ConditionalExecuteWithMpMemory::~ConditionalExecuteWithMpMemory() {
+    if (m_usingMpMemory) {
+        VOLT_DEBUG("Exiting UseMPmemory");
+        SynchronizedThreadLock::assumeLocalSiteContext();
+    }
+}
+
+ConditionalExecuteOutsideMpMemory::ConditionalExecuteOutsideMpMemory(bool haveMpMemory) : m_notUsingMpMemory(haveMpMemory) {
+    if (m_notUsingMpMemory) {
+        VOLT_DEBUG("Breaking out of UseMPmemory");
+        SynchronizedThreadLock::assumeLocalSiteContext();
+    }
+}
+
+ConditionalExecuteOutsideMpMemory::~ConditionalExecuteOutsideMpMemory() {
+    if (m_notUsingMpMemory) {
+        VOLT_DEBUG("Returning to UseMPmemory");
+        SynchronizedThreadLock::assumeMpMemoryContext();
+    }
+}
+
+ConditionalSynchronizedExecuteWithMpMemory::ConditionalSynchronizedExecuteWithMpMemory(bool needMpMemoryOnLowestThread,
+                                                                                       bool isLowestSite,
+                                                                                       int64_t& exceptionTracker) :
+        m_usingMpMemoryOnLowestThread(needMpMemoryOnLowestThread && isLowestSite),
+        m_okToExecute(!needMpMemoryOnLowestThread || m_usingMpMemoryOnLowestThread)
+{
+    if (needMpMemoryOnLowestThread) {
+        if (SynchronizedThreadLock::countDownGlobalTxnStartCount(isLowestSite)) {
+            // Call the execute method to actually perform whatever action
+            VOLT_DEBUG("Entering UseMPmemory");
+            SynchronizedThreadLock::assumeMpMemoryContext();
+            // Trap exceptions for replicated tables by initializing to an invalid value
+            exceptionTracker = -1;
+        }
+    }
+}
+ConditionalSynchronizedExecuteWithMpMemory::~ConditionalSynchronizedExecuteWithMpMemory() {
+    if (m_usingMpMemoryOnLowestThread) {
+        VOLT_DEBUG("Exiting UseMPmemory");
+        SynchronizedThreadLock::assumeLocalSiteContext();
+        SynchronizedThreadLock::signalLowestSiteFinished();
+    }
+}
+
+ExecuteWithAllSitesMemory::ExecuteWithAllSitesMemory()
+        : m_engineLocals()
+#ifndef NDEBUG
+        , m_wasUsingMpMemory(SynchronizedThreadLock::usingMpMemory())
+#endif
+{
+    assert(SynchronizedThreadLock::isLowestSiteContext());
+}
+
+ExecuteWithAllSitesMemory::~ExecuteWithAllSitesMemory()
+{
+    ExecutorContext::assignThreadLocals(m_engineLocals);
+#ifndef NDEBUG
+    SynchronizedThreadLock::setUsingMpMemory(m_wasUsingMpMemory);
+#endif
+}
+
+SharedEngineLocalsType::iterator ExecuteWithAllSitesMemory::begin() {
+    return SynchronizedThreadLock::s_enginesByPartitionId.begin();
+}
+
+SharedEngineLocalsType::iterator ExecuteWithAllSitesMemory::end() {
+    return SynchronizedThreadLock::s_enginesByPartitionId.end();
+}
+
+} // end namespace voltdb
