+++ tools/VoltDBCompilation.cmake
@@ -44,10 +44,10 @@ FUNCTION (VOLTDB_ADD_EXECUTABLE NAME)
 ENDFUNCTION()
 
 IF (IS_VALGRIND_BUILD)
-  VOLTDB_ADD_COMPILE_OPTIONS(-g3 -DDEBUG -DMEMCHECK )
+  VOLTDB_ADD_COMPILE_OPTIONS(-g3 -DMEMCHECK)
   SET (VOLTDB_USE_VALGRIND --valgrind)
 ELSEIF (VOLTDB_BUILD_TYPE STREQUAL "DEBUG")
-  VOLTDB_ADD_COMPILE_OPTIONS(-g3 -DDEBUG)
+  VOLTDB_ADD_COMPILE_OPTIONS(-g3)
 ELSEIF (VOLTDB_BUILD_TYPE STREQUAL "RELEASE")
   VOLTDB_ADD_COMPILE_OPTIONS(-O3 -g3 -mmmx -msse -msse2 -msse3 -DNDEBUG)
 ELSE()
